add listing of completed tasks
This adds a `list-closed` option for printing predictions which have
associated outcomes. This necessitated rewriting some of the
functionality for printing open predictions to make it more general.

The listing of completed predictions is not yet finished, as I need to
add the outcome and Brier score (based on the most recent forecast).